refactor: 수익률 계산 메소드 LottoResult 로 이동
